Crate axplat

axplat

This crate defines unified interfaces for various hardware platforms.

Modules

console

Console input and output.

init

Platform initialization.

irq

Interrupt request (IRQ) handling.

mem

Physical memory information.

power

Power management.

time

Time-related operations.

Macros

assert_str_eq

Checks that two strings are equal. If they are not equal, it will cause a compile-time error. And the message will be printed if it is provided.

console_print

Simple console print operation.

console_println

Simple console print operation, with a newline.

Functions

call_main

Call the function decorated by axplat::main for the primary core.

call_secondary_main

Call the function decorated by axplat::secondary_main for secondary cores.

Attribute Macros

impl_plat_interface

Implement the interface for a struct.

main

Marks a function to be called on the primary core after the platform initialization.

secondary_main

Marks a function to be called on the secondary cores after the platform initialization.