axplat

Attribute Macro impl_plat_interface

#[impl_plat_interface]

Implement the interface for a struct.

This attribute should be added above the implementation of a trait for a struct, and the trait must be defined with #[def_interface].

It is not necessary to implement it in the same crate as the definition, but it is required that these crates are linked together.

See the crate-level documentation for more details.