axplat

Module mem

Physical memory information.

Structs

MemRegionFlags

The flags of a physical memory region.

PhysMemRegion

A physical memory region.

Constants

DEFAULT_MMIO_FLAGS

The default flags for a MMIO region (readable, writable, device, and reserved).

DEFAULT_RAM_FLAGS

The default flags for a normal memory region (readable, writable and allocatable).

DEFAULT_RESERVED_FLAGS

The default flags for a reserved memory region (readable, writable, and reserved).

Traits

MemIf

Physical memory interface.

Functions

check_sorted_ranges_overlap

Checks if the given ranges are overlapping.

mmio_ranges

Returns all device memory (MMIO) ranges on the platform.

phys_ram_ranges

Returns all physical memory (RAM) ranges on the platform.

ranges_difference

Removes a portion of ranges from the given ranges.

reserved_phys_ram_ranges

Returns all reserved physical memory ranges on the platform.

total_ram_size

Returns the total size of physical memory (RAM) on the platform.

Type Aliases

OverlapErr

The error type for overlapping check.

RawRange

The raw memory range with start and size.